requirements.txt
bdateutil==0.1
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.11.1
django-allauth==0.32.0
django-appconf==1.0.2
django-braces==1.11.0
django-countries==4.5
django-crispy-forms==1.6.1
django-filer==1.2.7
django-import-export==0.5.1
django-jet==1.0.6
django-keyedcache3==1.5.3
django-measurement==2.4.0
django-model-utils==3.0.0
django-modeltranslation==0.12.1
django-money==0.10.2
django-mptt==0.8.7
django-polymorphic==1.0.2
django-registration-redux==1.6
django-rosetta==0.7.13
django-storages==1.5.2
django-taggit==0.22.1
django-tinymce==2.6.0
djangorestframework==3.6.3
djangorestframework-recursive==0.1.1
easy-thumbnails==2.4.1
google-api-python-client==1.4.1
gunicorn==19.7.1
holidays==0.8.1
httplib2==0.10.3
measurement==1.8.0
microsofttranslator==0.8
mysqlclient==1.3.10
oauth2client==4.1.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
polib==1.0.8
py-moneyed==0.7.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
python-dateutil==2.2
python3-openid==3.1.0
pytz==2017.2
requests==2.14.2
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.10.0
sympy==0.7.6.1
tablib==0.11.4
Unidecode==0.4.20
uritemplate==3.0.0
urllib3==1.21.1
whitenoise==3.3.0
